add raw Search input and refactor
