--- client/src/components/ExploreJobs/style.jsx
@@ -1,7 +1,0 @@
-import styled from "styled-components";
-
-export const Heading = styled.h2`
-  padding: 0.5em 0;
-  font-weight: 700;
-  font-size: 36px;
-`;
