--- client/src/components/pages/ExploreJobs.js
@@ -1,8 +1,0 @@
-import * as React from "react";
-import { Heading } from "../ExploreJobs/style";
-
-function ExploreJobs(props) {
-  return [<Heading key={0}>Find your next job</Heading>];
-}
-
-export default ExploreJobs;
