+++ client/src/components/pages/ExploreJobs/index.js
@@ -1,0 +1,2 @@
+import ExploreJobs from "./ExploreJobs";
+export default ExploreJobs;
